requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
packaging==23.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
requests==2.28.2
sqlparse==0.4.3
configparser==5.3.0

